requirements.txt
pytest==9.0.1
langchain==1.2.0
langchain-openai==1.1.6
python-dotenv==1.2.1
beautifulsoup4==4.14.3
langchain-community==0.4.1